1_build_libcorefoundation.sh: adopted to FreeBSD (without CFNetwork).

+++ Packaging/Sources/1_build_libcorefoundation.sh
@@ -6,8 +6,8 @@
 #----------------------------------------
 ${ECHO} ">>> Installing ${OS_ID} packages for CoreFoundation library build"
 if [ ${OS_ID} = "freebsd" ]; then
-	${ECHO} "FreeBSD: calling 'pkg install'."
-	sudo pkg install ${CORE_SYSTEM_DEPS}
+	${ECHO} "FreeBSD: calling 'pkg install'..."
+	sudo pkg install --yes --quiet ${CORE_SYSTEM_DEPS}
 elif [ ${OS_ID} = "debian" ] || [ ${OS_ID} = "ubuntu" ]; then
 	${ECHO} "Debian-based Linux distribution: calling 'apt-get install'."
 	sudo apt-get install -y ${RUNTIME_DEPS} || exit 1
@@ -44,7 +44,7 @@ C_FLAGS="-I/usr/local/include -I/usr/NextSpace/include -Wno-switch -Wno-enum-con
 $CMAKE_CMD .. \
 	-DCMAKE_C_COMPILER=${C_COMPILER} \
 	-DCMAKE_C_FLAGS="${C_FLAGS}" \
-	-DCMAKE_SHARED_LINKER_FLAGS="-L/usr/NextSpace/lib -luuid" \
+	-DCMAKE_SHARED_LINKER_FLAGS="-L/usr/local/lib -L/usr/NextSpace/lib -luuid" \
 	-DCF_DEPLOYMENT_SWIFT=NO \
 	-DBUILD_SHARED_LIBS=YES \
 	-DCMAKE_INSTALL_PREFIX=/usr/NextSpace \
@@ -58,28 +58,29 @@ $CMAKE_CMD .. \
 $MAKE_CMD || exit 1
 
 # CFNetwork
-if [ -n  "$libcfnetwork_version" ]; then
-	cd ${BUILD_ROOT}/${CFNET_PKG_NAME} || exit 1
-	rm -rf .build 2>/dev/null
-	mkdir -p .build
-	cd .build
-	CFN_CFLAGS="-F../../${CF_PKG_NAME}/.build -I/usr/local/include -I/usr/NextSpace/include"
-	CFN_LD_FLAGS="-L/usr/NextSpace/lib -L../../${CF_PKG_NAME}/.build/CoreFoundation.framework"
-	cmake .. \
-		-DCMAKE_C_COMPILER=${C_COMPILER} \
-		-DCMAKE_CXX_COMPILER=clang++ \
-		-DCFNETWORK_CFLAGS="${CFN_CFLAGS}" \
-		-DCFNETWORK_LDLAGS="${CFN_LD_FLAGS}" \
-		-DBUILD_SHARED_LIBS=YES \
-		-DCMAKE_INSTALL_PREFIX=/usr/NextSpace \
-		-DCMAKE_INSTALL_LIBDIR=/usr/NextSpace/lib \
-		\
-		-DCMAKE_SKIP_RPATH=ON \
-		-DCMAKE_BUILD_TYPE=Debug
-	$MAKE_CMD || exit 1
+# No support for FreeBSD yet
+if [ ${OS_ID} != "freebsd" ]; then
+	if [ -n  "$libcfnetwork_version" ]; then
+		cd ${BUILD_ROOT}/${CFNET_PKG_NAME} || exit 1
+		rm -rf .build 2>/dev/null
+		mkdir -p .build
+		cd .build
+		CFN_CFLAGS="-F../../${CF_PKG_NAME}/.build -I/usr/local/include -I/usr/NextSpace/include"
+		CFN_LD_FLAGS="-L/usr/NextSpace/lib -L../../${CF_PKG_NAME}/.build/CoreFoundation.framework"
+		cmake .. \
+			-DCMAKE_C_COMPILER=${C_COMPILER} \
+			-DCMAKE_CXX_COMPILER=clang++ \
+			-DCFNETWORK_CFLAGS="${CFN_CFLAGS}" \
+			-DCFNETWORK_LDLAGS="${CFN_LD_FLAGS}" \
+			-DBUILD_SHARED_LIBS=YES \
+			-DCMAKE_INSTALL_PREFIX=/usr/NextSpace \
+			-DCMAKE_INSTALL_LIBDIR=/usr/NextSpace/lib \
+			\
+			-DCMAKE_SKIP_RPATH=ON \
+			-DCMAKE_BUILD_TYPE=Debug
+		$MAKE_CMD || exit 1
+	fi
 fi
-
-exit 0
 
 #----------------------------------------
 # Install
@@ -108,27 +109,29 @@ $RM_CMD libCoreFoundation.so*
 $LN_CMD ../Frameworks/CoreFoundation.framework/Versions/${libcorefoundation_version}/libCoreFoundation.so* ./
 
 ### CFNetwork
-if [ -n  "$libcfnetwork_version" ]; then
-	cd ${BUILD_ROOT}/${CFNET_PKG_NAME}/.build || exit 1
-	$INSTALL_CMD
+if [ ${OS_ID} != "freebsd" ]; then
+	if [ -n  "$libcfnetwork_version" ]; then
+		cd ${BUILD_ROOT}/${CFNET_PKG_NAME}/.build || exit 1
+		$INSTALL_CMD
 
-	CFNET_DIR=${DEST_DIR}/usr/NextSpace/Frameworks/CFNetwork.framework
+		CFNET_DIR=${DEST_DIR}/usr/NextSpace/Frameworks/CFNetwork.framework
 
-	$MKDIR_CMD $CFNET_DIR/Versions/${libcfnetwork_version}
-	cd $CFNET_DIR
-	# Headers
-	$MV_CMD Headers Versions/${libcfnetwork_version}
-	$LN_CMD Versions/Current/Headers Headers
-	cd Versions
-	$LN_CMD ${libcfnetwork_version} Current
-	cd ..
-	# Libraries
-	$MV_CMD libCFNetwork.so* Versions/${libcfnetwork_version}
-	$LN_CMD Versions/Current/libCFNetwork.so.${libcfnetwork_version} libCFNetwork.so
-	$LN_CMD Versions/Current/libCFNetwork.so.${libcfnetwork_version} CFNetwork
-	cd ../../lib
-	$RM_CMD libCFNetwork.so*
-	$LN_CMD ../Frameworks/CFNetwork.framework/Versions/${libcfnetwork_version}/libCFNetwork.so* ./
+		$MKDIR_CMD $CFNET_DIR/Versions/${libcfnetwork_version}
+		cd $CFNET_DIR
+		# Headers
+		$MV_CMD Headers Versions/${libcfnetwork_version}
+		$LN_CMD Versions/Current/Headers Headers
+		cd Versions
+		$LN_CMD ${libcfnetwork_version} Current
+		cd ..
+		# Libraries
+		$MV_CMD libCFNetwork.so* Versions/${libcfnetwork_version}
+		$LN_CMD Versions/Current/libCFNetwork.so.${libcfnetwork_version} libCFNetwork.so
+		$LN_CMD Versions/Current/libCFNetwork.so.${libcfnetwork_version} CFNetwork
+		cd ../../lib
+		$RM_CMD libCFNetwork.so*
+		$LN_CMD ../Frameworks/CFNetwork.framework/Versions/${libcfnetwork_version}/libCFNetwork.so* ./
+	fi
 fi
 
 if [ "$DEST_DIR" = "" ]; then
